Add version dependent inclusion of argparse as a install requirement.
For versions prior to 2.7 argparse must be installed manually before
mpyq can be installed for two reasons this patch addresses:

    1. __version__ is imported from mpyq.py which requires all the
       dependencies to already be installed before the version can
       be imported.

    2. The setup.py file doesn't include a install_requires keyword
       in the call to the setup functions.

By manually specifying the version in the setup.py file we avoid the
chicken and the egg problem. By checking the current python version
we can make sure to only require the argparse module for versions of
python < 2.7.